book-search-bedrock/requirements.txt
boto3
opensearch-py
python-dotenv
numpy
streamlit